feat: rework attachments, expose to templates, fix audit persistence
Replace the two-variant Attachment enum (BlobRef/Inline) with a flat
struct carrying id, name, filename, content_type, and data_base64.
Remove the acteon-blob crate entirely. Expose attachment metadata in
the template engine context (attachments list + attachments_by_id map).
Fix audit trail persistence — all backends (Postgres, ClickHouse,
DynamoDB) now store and retrieve attachment_metadata. Fix size_bytes
to reflect decoded binary size instead of base64 string length.

Extract gateway.rs into audit_helpers, quota_enforcement, and
template_management modules (~730 lines moved out).
