--- crates/blob/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "acteon-blob"
-description = "Pluggable blob storage trait for Acteon attachment support"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-async-trait = { workspace = true }
-bytes = { workspace = true }
-chrono = { workspace = true }
-serde = { workspace = true }
-thiserror = { workspace = true }
-uuid = { workspace = true }
-
-[lints]
-workspace = true
