--- crates/blob/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod error;
-pub mod store;
-pub mod types;
-
-pub use error::BlobError;
-pub use store::BlobStore;
-pub use types::{BlobMetadata, ResolvedBlob};
