+++ crates/gateway/src/builder.rs
@@ -65,7 +65,6 @@ pub struct GatewayBuilder {
     resource_lookups: HashMap<String, Arc<dyn ResourceLookup>>,
     templates: HashMap<(String, String), HashMap<String, acteon_core::Template>>,
     template_profiles: HashMap<(String, String), HashMap<String, acteon_core::TemplateProfile>>,
-    blob_store: Option<Arc<dyn acteon_blob::BlobStore>>,
     max_inline_bytes: u64,
     max_attachments_per_action: usize,
 }
@@ -110,7 +109,6 @@ impl GatewayBuilder {
             resource_lookups: HashMap::new(),
             templates: HashMap::new(),
             template_profiles: HashMap::new(),
-            blob_store: None,
             max_inline_bytes: 5_242_880,
             max_attachments_per_action: 10,
         }
@@ -487,18 +485,7 @@ impl GatewayBuilder {
         self
     }
 
-    /// Set the pluggable blob store for resolving `BlobRef` attachments.
-    ///
-    /// When set, actions with `BlobRef` attachments can resolve their blob data
-    /// at dispatch time. Without a blob store, only inline base64 attachments
-    /// are supported.
-    #[must_use]
-    pub fn blob_store(mut self, store: Arc<dyn acteon_blob::BlobStore>) -> Self {
-        self.blob_store = Some(store);
-        self
-    }
-
-    /// Set the maximum allowed size for inline base64 attachments (default: 5 MB).
+    /// Set the maximum allowed size for a single attachment after `base64` decoding (default: 5 MB).
     #[must_use]
     pub fn max_inline_bytes(mut self, max: u64) -> Self {
         self.max_inline_bytes = max;
@@ -790,7 +777,6 @@ impl GatewayBuilder {
             resource_lookups: self.resource_lookups,
             templates: parking_lot::RwLock::new(self.templates),
             template_profiles: parking_lot::RwLock::new(self.template_profiles),
-            blob_store: self.blob_store,
             max_inline_bytes: self.max_inline_bytes,
             max_attachments_per_action: self.max_attachments_per_action,
         })
