+++ crates/integrations/webhook/src/provider.rs
@@ -226,13 +226,13 @@ impl Provider for WebhookProvider {
 
         let mut form = reqwest::multipart::Form::new().text("payload", body_json.clone());
 
-        for (i, blob) in ctx.attachments.iter().enumerate() {
-            let part = reqwest::multipart::Part::bytes(blob.data.to_vec())
-                .file_name(blob.metadata.filename.clone())
-                .mime_str(&blob.metadata.content_type)
+        for (i, resolved) in ctx.attachments.iter().enumerate() {
+            let part = reqwest::multipart::Part::bytes(resolved.data.clone())
+                .file_name(resolved.filename.clone())
+                .mime_str(&resolved.content_type)
                 .unwrap_or_else(|_| {
-                    reqwest::multipart::Part::bytes(blob.data.to_vec())
-                        .file_name(blob.metadata.filename.clone())
+                    reqwest::multipart::Part::bytes(resolved.data.clone())
+                        .file_name(resolved.filename.clone())
                 });
             form = form.part(format!("file_{i}"), part);
         }
