+++ ui/src/pages/Dispatch.module.css
@@ -135,3 +135,21 @@
     color: var(--text-default);
   }
 }
+
+.attachmentFields {
+  display: flex;
+  gap: 0.5rem;
+  flex: 1;
+  min-width: 0;
+}
+
+.attachmentInput {
+  flex: 1;
+  min-width: 0;
+  padding: 0.125rem 0.375rem;
+  font-size: 0.75rem;
+  border: 1px solid var(--color-border);
+  border-radius: 0.25rem;
+  background: var(--color-bg);
+  color: var(--color-text);
+}
